define TEMP_FOLDER environment variable

+++ server/system/schedulers/tempFilesCleanup.js
@@ -6,7 +6,7 @@ const Logger = require('../../log/log').getLogger('TempFilesCleanup')
 
 const DateUtils = require('../../../common/dateUtils')
 
-const tempFolder = '/tmp/arena-upload'
+const Environment = require('../environment')
 
 const initSchedule = () =>
   // execute the cron job every day at 2AM
@@ -16,35 +16,31 @@ const initSchedule = () =>
   })
 
 const cleanupTempFiles = async (olderThanHours = null) => {
-  Logger.debug('Cleaning up temp files')
+  const tempFolder = Environment.tempFolder
+
+  Logger.debug(`Cleaning up temp files in folder ${tempFolder}`)
 
   let count = 0
-  const now = new Date()
-
-  await visitTempFiles(async (filePath, stat) => {
-    //delete temp files older than 'olderThanHours'
-    if (!olderThanHours || DateUtils.differenceInHours(now, new Date(stat.ctime)) >= olderThanHours) {
-      await fs.unlinkSync(filePath)
-      Logger.debug('Temp file deleted', filePath)
-      count++
-    }
-  })
-
-  Logger.debug(`${count} temp files deleted`)
-}
-
-const visitTempFiles = async visitor => {
   try {
-    const files = await fs.readdirSync(tempFolder)
-    for (const file of files) {
-      const filePath = path.join(tempFolder, file)
-      const stat = await fs.statSync(filePath)
-      if (stat.isFile())
-        await visitor(filePath, stat)
+    if (await fs.existsSync(tempFolder)) {
+      const now = new Date()
+      const files = await fs.readdirSync(tempFolder)
+      for (const file of files) {
+        const filePath = path.join(tempFolder, file)
+        const stat = await fs.statSync(filePath)
+        if (stat.isFile() && (!olderThanHours || DateUtils.differenceInHours(now, new Date(stat.ctime)) >= olderThanHours)) {
+          await fs.unlinkSync(filePath)
+          Logger.debug('Temp file deleted', filePath)
+          count++
+        }
+      }
     }
   } catch (error) {
-    //ignore errors (eg. temp folder not created yet, unable to delete temp file, etc)
+    //ignore errors
+    Logger.error('Error deleting temp files', error)
   }
+
+  Logger.debug(`${count} temp files deleted`)
 }
 
 const init = async () => {
@@ -54,7 +50,5 @@ const init = async () => {
 }
 
 module.exports = {
-  tempFolder,
-
   init
 }
